removes old commented code and moves ALL_SYMBOLS t StockServices
