+++ client/src/StocksInfoComponents/StockChart.js
@@ -1,7 +1,7 @@
 import Highcharts from 'highcharts/highstock'
 import HighchartsReact from 'highcharts-react-official'
 
-const StockChart = ({IBM, AAPL, MSFT, TSLA, XOM, WMT, NVDA, META, dailyStockInfo, comparePercent}) => {
+const StockChart = ({dailyStockInfo, comparePercent}) => {
     
     const stockSeries = dailyStockInfo.map((stockData) => {
         
@@ -49,42 +49,7 @@ const StockChart = ({IBM, AAPL, MSFT, TSLA, XOM, WMT, NVDA, META, dailyStockInfo
             verticalAlign: 'middle',
             padding: 40
         },
-        // plotOptions: undefined,
         series: stockSeries
-        // [{
-        //     data: IBM,
-        //     name: 'IBM',
-
-        // },
-        // {
-        //     data: AAPL,
-        //     name: 'AAPL'
-        // },
-        // {
-        //     data: MSFT,
-        //     name: 'MSFT'
-        // },
-        // {
-        //     data: TSLA,
-        //     name: 'TSLA'
-        // },
-        // {
-        //     data: XOM,
-        //     name: 'XOM'
-        // },
-        // {
-        //     data: NVDA,
-        //     name: 'NVDA'
-        // },
-        // {
-        //     data: META,
-        //     name: 'META'
-        // },
-        // {
-        //     data: WMT,
-        //     name: 'WMT'
-        // },
-        // ]
     };
 
     comparePercent ? options.plotOptions = {series: {compare: 'percent'}} : options.plotOptions = {series: {compare: undefined}}
