+++ client/src/containers/StocksInfoContainer.js
@@ -3,7 +3,7 @@ import { getDailyBySymbol } from "../services/StockServices"
 import styled from "styled-components"
 import StockChart from "../StocksInfoComponents/StockChart"
 import SwitchYButton from "../StocksInfoComponents/SwitchYButton"
-import { ALL_SYMBOLS } from "../services/TickerService"
+import { ALL_SYMBOLS } from "../services/StockServices"
 
 
 
@@ -12,16 +12,6 @@ const StocksInfoContainer = () => {
     const [comparePercent, setComparePercent] = useState(true)
 
     const [dailyStockInfo, setDailyStockInfo] = useState([])
-    
-    // const [IBMDaily, setIBMDaily] = useState([])
-    // const [AAPLDaily, setAAPLDaily] = useState([])
-    // const [MSFTDaily, setMSFTDaily] = useState([])
-    // const [TSLADaily, setTSLADaily] = useState([])
-    // const [XOMDaily, setXOMDaily] = useState([])
-    // const [WMTDaily, setWMTDaily] = useState([])
-    // const [NVDADaily, setNVDADaily] = useState([])
-    // const [METADaily, setMETADaily] = useState([])
-
 
     useEffect( () => {
 
@@ -30,39 +20,6 @@ const StocksInfoContainer = () => {
         }))
 
         Promise.all(stockInfoPromises).then(data => setDailyStockInfo(data))
-
-        // getDailyBySymbol('IBM')
-        //     .then(allDaily => {
-        //         setIBMDaily(allDaily)
-        //     })
-        // getDailyBySymbol('AAPL')
-        //     .then(allDaily => {
-        //         setAAPLDaily(allDaily)
-        //     })
-        // getDailyBySymbol('MSFT')
-        //     .then(allDaily => {
-        //         setMSFTDaily(allDaily)         
-        //     })
-        // getDailyBySymbol('TSLA')
-        //     .then(allDaily => {
-        //         setTSLADaily(allDaily)         
-        //     })
-        // getDailyBySymbol('XOM')
-        //     .then(allDaily => {
-        //         setXOMDaily(allDaily)   
-        //     })
-        // getDailyBySymbol('WMT')
-        //     .then(allDaily => {
-        //         setWMTDaily(allDaily)
-        //     })
-        // getDailyBySymbol('NVDA')
-        //     .then(allDaily => {
-        //         setNVDADaily(allDaily)        
-        //     })
-        // getDailyBySymbol('META')
-        //     .then(allDaily => {
-        //         setMETADaily(allDaily)
-        //     })
 
     }, [])
 
@@ -77,14 +34,6 @@ const StocksInfoContainer = () => {
     <Wrapper>
         <StockChart 
             dailyStockInfo={dailyStockInfo}
-            // IBM={IBMDaily} 
-            // AAPL={AAPLDaily} 
-            // MSFT={MSFTDaily}
-            // TSLA={TSLADaily}
-            // XOM={XOMDaily}
-            // WMT={WMTDaily}
-            // NVDA={NVDADaily}
-            // META={METADaily}
             comparePercent={comparePercent}/>
         <SwitchYButton flipComparePercent={flipComparePercent}/>
     </Wrapper>
